make recordscreen directory
