+++ app/src/main/java/com/example/Healsenior/recordScreen/RecordScreenHeader.kt
@@ -1,0 +1,57 @@
+package com.example.Healsenior.recordScreen
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.Healsenior.R
+
+@Composable
+fun RecordScreenHeader() {
+    Row(modifier = Modifier
+        .fillMaxWidth()
+        .height(60.dp)
+        .background(color = Color(0xFF95BDFA))
+    )
+    {
+        Row(
+            modifier = Modifier.weight(1f),
+            horizontalArrangement = Arrangement.Start
+        ) {
+            Text(
+                text = "기록",
+                fontWeight = FontWeight.Bold,
+                fontSize = 30.sp,
+                modifier = Modifier
+                    .padding(start = 20.dp, top = 15.dp)
+            )
+        }
+        Row(
+            modifier = Modifier.weight(1f),
+            horizontalArrangement = Arrangement.End
+        ) {
+            val image = painterResource(R.drawable.img1)
+
+            Image(
+                painter = image,
+                contentDescription = null,
+                modifier = Modifier
+                    .width(50.dp)
+                    .height(50.dp)
+                    .padding(top = 15.dp, end = 15.dp)
+            )
+        }
+    }
+}
